requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
lazy-object-proxy==1.4.3
lxml==4.6.1
mccabe==0.6.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
urllib3==1.25.11
wrapt==1.12.1
mysql-connector-python-rf==2.2.2